+++ TZMate/App/AppDelegate.swift
@@ -1,3 +1,10 @@
+//
+//  TZ Mate
+//  Copyright (c) 2026 Anton Pavlov
+//  GitHub: https://github.com/RemDee13
+//  Licensed under the MIT License.
+//
+
 import AppKit
 import SwiftUI
 
@@ -27,7 +34,7 @@ final class AppDelegate: NSObject, NSApplicationDelegate, NSPopoverDelegate {
             button.title = "TZ"
             button.toolTip = "TZ Mate"
             button.target = self
-            button.action = #selector(togglePopover(_:))
+            button.action = #selector(handleStatusItemClick(_:))
             button.sendAction(on: [.leftMouseUp, .rightMouseUp])
         }
 
@@ -50,12 +57,58 @@ final class AppDelegate: NSObject, NSApplicationDelegate, NSPopoverDelegate {
     }
 
     @objc
+    private func handleStatusItemClick(_ sender: AnyObject?) {
+        if shouldShowStatusMenu {
+            showStatusMenu()
+            return
+        }
+
+        togglePopover(sender)
+    }
+
+    @objc
+    private func openFromStatusMenu(_ sender: AnyObject?) {
+        showPopover()
+    }
+
+    @objc
+    private func quitFromStatusMenu(_ sender: AnyObject?) {
+        popover.close()
+        NSApp.terminate(nil)
+    }
+
     private func togglePopover(_ sender: AnyObject?) {
         if popover.isShown {
             popover.performClose(sender)
         } else {
             showPopover()
         }
+    }
+
+    private var shouldShowStatusMenu: Bool {
+        guard let event = NSApp.currentEvent else {
+            return false
+        }
+
+        return event.type == .rightMouseUp || event.modifierFlags.contains(.control)
+    }
+
+    private func showStatusMenu() {
+        let menu = NSMenu()
+
+        let openItem = NSMenuItem(title: "Open TZ Mate", action: #selector(openFromStatusMenu(_:)), keyEquivalent: "")
+        openItem.target = self
+        menu.addItem(openItem)
+
+        menu.addItem(.separator())
+
+        let quitItem = NSMenuItem(title: "Quit TZ Mate", action: #selector(quitFromStatusMenu(_:)), keyEquivalent: "q")
+        quitItem.target = self
+        menu.addItem(quitItem)
+
+        statusItem?.menu = menu
+        statusItem?.button?.performClick(nil)
+        statusItem?.menu = nil
     }
 
     private func showPopover() {